dist/datachannel_test_2peer/send/001.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Send Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div id="signallingInfo"></div>
		<hr />
		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
			setup({
                explicit_timeout : true
            });
            var label = "test";
            var dataChannel = null;
            var expected = "s", size = 14, repeats = 100, messageCount = 0;
            for (var i = 0; i < size; i++) {
                expected += expected;
            }
            var test = async_test("Set up a DataChannel - send multiple data packets - size = " + expected.length / 1024 + " KB, repeats = " + repeats, {timeout: 180000});
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {
                        for (var i = 0; i < repeats; i++) {
                            dataChannel.send(expected);
                        }
                        test.done();
                    });
                });
            } else {
            	var complete = 0;
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.onmessage = test.step_func(function(e) {
                        if ((e.data).length == expected.length)
                            messageCount++;
                        if (messageCount == repeats) {
                            assert_equals(e.data, expected, "Received data not correct ");
                            test.done();
                        }
                    });
                });
            }
		</script>
	</body>
</html>

